Split remaining validators into partial classes
